Add UDPC_set_con_timeout_millis(...)
Add new function to set the connection timeout time. By default, this
timeout time is 10 seconds. If a packet hasn't been received for this
duration, then the connection is considered "timed out" and is dropped.

Previously this time was a constant 10 seconds. This commit allows
changing the connection timeout time. However, it cannot be less than
the heartbeat interval time.