TSLQueue: Use atomic int for better iterators
The previous change of removing "trade read-for-write lock" puts
iterators in an inconsistent state if the trade is required.

An iterator instance starts its existence with a read lock. If changes
need to be made (like removing an element in the iterator), it needs a
write lock. Previously, the implementation had the iterator drop a read
lock to obtain a write lock. The newly added atomic int keeps track if
there exists iterators, and protects against obtaining a write lock from
something other than the iterator. In other words, the TSLQueue can only
obtain a write lock if there are no iterators. This also allows the
iterator's read-for-write lock swap to be more robust as there is no
risk of the underlying data to change during this swap (as the atomic
int protects against this).