Huge refactorings related to TSLQueue
Change how TSLQueue'd packets are iterated:
Instead of using an iterator, take the entire Queue (leaving an empty
queue) and work on that. If there are packets remaining, reinsert the
remaining into the left-behind-queue.

Added a C++ style struct to handle cleanup of UDPC_PacketInfo called
PktInfoWrapper.

Fixes and changes related to refactorings.